Remove unnecessary 'require' in the timestamp decoder
The commit 5eae85a changed the way the timezoneOffset was handled
and this require appears to have been an artifact of that change.

Instead of being an export from the vertica module, it is handled
directly by VerticaTimestamp.setTimezoneOffset().

Calling require() repeatly still hits filepath related code, even if
cached. That can be expensive.